Apply UINT32/INT32/UINT8 etc. type aliases across codebase
- Add types.h with UINT8/16/32/64, INT8/16/32/64, ErrNo aliases
- Replace all stdint types (uint32_t → UINT32, size_t → UINT32, etc.)
- Wire types.h into include chain via dsc_errors.h and dwarf_types.h
- bool stays as bool per coding style
- dwarf/ types also use new aliases (types are project-wide)

+++ projects/dsp-connect/src/arch/arch.h
@@ -5,6 +5,7 @@
 #ifndef DSC_ARCH_H
 #define DSC_ARCH_H
 
+#include "../util/types.h"
 #include <stddef.h>
 #include <stdint.h>
 
@@ -24,21 +25,21 @@ typedef struct {
 /* --- vtable: every backend implements these --- */
 struct dsc_arch_ops {
     /* Convert logical address (from DWARF) to physical address (for transport) */
-    int (*logical_to_physical)(const dsc_arch_t *self, uint64_t logical,
-                               uint64_t *physical);
+    int (*logical_to_physical)(const dsc_arch_t *self, UINT64 logical,
+                               UINT64 *physical);
 
     /* Convert physical address back to logical */
-    int (*physical_to_logical)(const dsc_arch_t *self, uint64_t physical,
-                               uint64_t *logical);
+    int (*physical_to_logical)(const dsc_arch_t *self, UINT64 physical,
+                               UINT64 *logical);
 
     /* Swap endianness of a value buffer (noop if host matches target) */
-    void (*swap_endian)(const dsc_arch_t *self, void *buf, size_t size);
+    void (*swap_endian)(const dsc_arch_t *self, void *buf, UINT32 size);
 
     /* Get minimum addressable unit size in bytes */
-    size_t (*min_access_size)(const dsc_arch_t *self);
+    UINT32 (*min_access_size)(const dsc_arch_t *self);
 
     /* Get word size in bytes */
-    size_t (*word_size)(const dsc_arch_t *self);
+    UINT32 (*word_size)(const dsc_arch_t *self);
 
     /* Destroy and free all resources */
     void (*destroy)(dsc_arch_t *self);
@@ -55,31 +56,31 @@ struct dsc_arch_t {
  * =================================================================== */
 
 static inline int dsc_arch_logical_to_physical(const dsc_arch_t *a,
-                                               uint64_t logical,
-                                               uint64_t *physical)
+                                               UINT64 logical,
+                                               UINT64 *physical)
 {
     return a->ops->logical_to_physical(a, logical, physical);
 }
 
 static inline int dsc_arch_physical_to_logical(const dsc_arch_t *a,
-                                               uint64_t physical,
-                                               uint64_t *logical)
+                                               UINT64 physical,
+                                               UINT64 *logical)
 {
     return a->ops->physical_to_logical(a, physical, logical);
 }
 
 static inline void dsc_arch_swap_endian(const dsc_arch_t *a,
-                                        void *buf, size_t size)
+                                        void *buf, UINT32 size)
 {
     a->ops->swap_endian(a, buf, size);
 }
 
-static inline size_t dsc_arch_min_access_size(const dsc_arch_t *a)
+static inline UINT32 dsc_arch_min_access_size(const dsc_arch_t *a)
 {
     return a->ops->min_access_size(a);
 }
 
-static inline size_t dsc_arch_word_size(const dsc_arch_t *a)
+static inline UINT32 dsc_arch_word_size(const dsc_arch_t *a)
 {
     return a->ops->word_size(a);
 }
